Merge development into main

See merge request cse1105/2022-2023/teams/oopp-team-42!65

+++ server/src/main/java/server/api/Card/CardService.java
@@ -1,13 +1,14 @@
 package server.api.Card;
 
-import commons.Result;
 import commons.Card;
+import commons.Result;
 import commons.Task;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import server.api.Task.TaskService;
 import server.database.CardRepository;
-import java.util.List;
+
+import java.util.*;
 
 /**
  * Handles logic of the Card endpoints
@@ -44,6 +45,15 @@ public class CardService {
             return Result.OBJECT_ISNULL.of(null);
         }
         try {
+            var tasks = card.taskList;
+            card.taskList = new ArrayList<>();
+            cardRepository.save(card);
+            for (var task :
+                    tasks) {
+                task.card = card;
+                taskService.addNewTask(task);
+            }
+            card.taskList = tasks;
             return Result.SUCCESS.of(cardRepository.save(card));
         }catch (Exception e){
             return Result.FAILED_ADD_NEW_CARD.of(null);
@@ -54,7 +64,7 @@ public class CardService {
      * Deletes the Card with the given id
      * @param id
      */
-    public Result<Object> deleteCard (Integer id) {
+    public Result<Object> deleteCard (UUID id) {
         try {
             cardRepository.deleteById(id);
             return Result.SUCCESS.of(null);
@@ -66,16 +76,20 @@ public class CardService {
     /**
      * Updates the name of the Card with specific id {id},
      * with the name of the given Card card.
-     *
      * @param card card with the new name
      * @param id id of the card to be updated
      */
-    public Result<Object> updateCard(Card card, Integer id){
+    public Result<Object> updateCard(Card card, UUID id){
 
         try {
             return Result.SUCCESS.of(cardRepository.findById(id)
                     .map(l -> {
                         l = card;
+                        for (var task :
+                                card.taskList) {
+                            task.card = card;
+                            taskService.updateTask(task, task.taskID);
+                        }
                         return cardRepository.save(l);
                     }));
         }catch (Exception e){
@@ -88,20 +102,20 @@ public class CardService {
      * @param id
      * @return card with specific ID
      */
-    public Result<Card> getCardById(Integer id){
-            try{
-                return Result.SUCCESS.of(cardRepository.findById(id).get());
-            }
-            catch (Exception e){
-                return Result.FAILED_RETRIEVE_CARD_BY_ID;
-            }
+    public Result<Card> getCardById(UUID id){
+        if(id == null) return Result.OBJECT_ISNULL.of(null);
+        try{
+            return Result.SUCCESS.of(cardRepository.findById(id).get());
+        } catch (Exception e){
+            return Result.FAILED_RETRIEVE_CARD_BY_ID;
+        }
     }
 
     /**
      * Removes a certain task from the card with Id {id}
      */
-    public Result<Card> removeTaskFromCard(Task task, Integer id){
-        //Remove task from repository
+    public Result<Card> removeTaskFromCard(Task task, UUID id){
+        if(task == null || id == null) return Result.OBJECT_ISNULL.of(null);
         try{
             taskService.deleteTask(task.taskID);
             return Result.SUCCESS.of(cardRepository.findById(id)
@@ -109,8 +123,7 @@ public class CardService {
                         c.taskList.remove(task);
                         return cardRepository.save(c);
                     }).get());
-        }
-        catch (Exception e){
+        } catch (Exception e){
             return Result.FAILED_REMOVE_CARD;
         }
     }
@@ -118,13 +131,14 @@ public class CardService {
     /**
      * Adds the given task to the card with Id {id}
      */
-    public Result<Card> addTaskToCard(Task task, Integer id){
+    public Result<Card> addTaskToCard(Task task, UUID id){
 //        Card card = cardRepository.findById(id).get();
 //        task.card = card;
 //        var result = taskService.addNewTask(task);
 //        if(!result.success) {
 //            return result.of(null);
 //        }
+        if(task == null || id == null) return Result.OBJECT_ISNULL.of(null);
         try{
             Card card = cardRepository.findById(id).get();
             task.card = card;
@@ -138,5 +152,12 @@ public class CardService {
         }
     }
 
-
+    /** Updates the card it receives in the database. */
+    public Result<Card> updateCard(Card card) {
+        try {
+            return Result.SUCCESS.of(cardRepository.save(card));
+        } catch (Exception e) {
+            return Result.FAILED_UPDATE_CARD.of(null);
+        }
+    }
 }
