+++ server/src/main/java/server/api/List/ListService.java
@@ -5,10 +5,12 @@ import commons.CardList;
 import commons.Result;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import server.api.Board.BoardService;
 import server.api.Card.CardService;
 import server.database.ListRepository;
 
 import java.util.List;
+import java.util.UUID;
 
 /**
  * Handles business logic of the List endpoints
@@ -18,9 +20,12 @@ public class ListService {
 
     private final ListRepository listRepository;
     private final CardService cardService;
+    private final BoardService boardService;
 
+    /** Initialises the controller using dependency injection */
     @Autowired
-    public ListService(ListRepository listRepository, CardService cardService) {
+    public ListService(ListRepository listRepository, CardService cardService, BoardService boardService) {
+        this.boardService = boardService;
         this.listRepository = listRepository;
         this.cardService = cardService;
     }
@@ -44,7 +49,9 @@ public class ListService {
             return Result.OBJECT_ISNULL.of(null);
         }
         try {
-            return Result.SUCCESS.of(listRepository.save(list));
+            var resultList = listRepository.save(list);
+            boardService.updateBoardAddList(list);
+            return Result.SUCCESS.of(resultList);
         }catch (Exception e){
             return Result.FAILED_ADD_NEW_LIST;
         }
@@ -53,21 +60,24 @@ public class ListService {
     /**
      * Deletes the CardList with the given id
      */
-    public Result<Object> deleteList(Integer id) {
+    public Result<Object> deleteList(UUID id) {
+        if(id == null) return Result.OBJECT_ISNULL.of(null);
         try {
+            var l =listRepository.findById(id);
+            if(l.isEmpty()) return Result.FAILED_DELETE_LIST.of(null);
+            boardService.deleteList(l.get());
             listRepository.deleteById(id);
-            return Result.SUCCESS;
+            return Result.SUCCESS.of(true);
         }catch (Exception e){
-            return Result.FAILED_DELETE_LIST;
+            return Result.FAILED_DELETE_LIST.of(false);
         }
     }
-
 
     /**
      * Updates the name of the CardList with id {id},
      * with the name of the given CardList list.
      */
-    public Result<CardList> updateName(CardList list, Integer id) {
+    public Result<CardList> updateName(CardList list, UUID id) {
         try {
             return Result.SUCCESS.of(listRepository.findById(id)
                     .map(l -> {
@@ -82,7 +92,7 @@ public class ListService {
     /**
      * Get a list by an id method
      */
-    public Result<CardList> getListById(Integer id) {
+    public Result<CardList> getListById(UUID id) {
         try {
             return Result.SUCCESS.of(listRepository.findById(id).get());
         }catch (Exception e){
@@ -93,7 +103,7 @@ public class ListService {
     /**
      * Removes a certain card from the list with Id {id}
      */
-    public Result<CardList> removeCardFromList(Card card, Integer id){
+    public Result<CardList> removeCardFromList(Card card, UUID id){
         try{
             cardService.deleteCard(card.cardID);
             return Result.SUCCESS.of(listRepository.findById(id)
@@ -105,41 +115,29 @@ public class ListService {
         } catch (Exception e){
             return Result.FAILED_REMOVE_CARD_FROM_LIST;
         }
-//
-//        cardService.deleteCard(card.cardID);
-//        return Result.SUCCESS.of(listRepository.findById(id)
-//                .map(l -> {
-//                    l.cardList.remove(card);
-//                    return listRepository.save(l);
-//                }).get());
     }
 
-    /**
-     * Adds the given card to the list with Id {id}
+
+    /** Adds a card to a list in the repo
+     * @param card
+     * @param listId
+     * @return
      */
-    public Result<Card> addCardToList(Card card, Integer id){
+    public Result<Card> addCardToList(Card card, UUID listId){
         try{
             Result<Card> result = cardService.addNewCard(card);
-            return Result.SUCCESS.of(listRepository.findById(id)
+            return Result.SUCCESS.of(listRepository.findById(listId)
                     .map(l -> {
                         if(!l.cardList.contains(card)){
                             l.cardList.add(card);
                         }
+                        System.out.println("List with id\t" + listId + "\thas size\t" + l.cardList.size());
                         listRepository.save(l);
                         return result.value;
                     }).get());
         } catch (Exception e){
             return Result.OBJECT_ISNULL.of(null);
         }
-//        var list = listRepository.findById(id).get();
-//        card.cardList = list;
-//        var result = cardService.addNewCard(card);
-//        if (!result.success) {
-//            return result.of(null);
-//        }
-//        list.addCard(card);
-//        listRepository.save(list);
-//        return Result.SUCCESS.of(card);
     }
 
     /**
@@ -154,7 +152,7 @@ public class ListService {
      * as there is a bidirectional reference between the card and the cardList.
      *
      */
-    public Result<Card> moveCard(Card card, Integer idFrom, Integer idTo,int indexTo) {
+    public Result<Card> moveCard(Card card, UUID idFrom, UUID idTo,int indexTo) {
         if(card == null || idFrom == null || idTo == null) {
             return Result.OBJECT_ISNULL.of(null);
         }
@@ -165,11 +163,12 @@ public class ListService {
             oldCardList.cardList.remove(card);
 
             card.cardList = newCardList;
-            card.cardListId = newCardList.cardListID;
+            card.cardListId = newCardList.cardListId;
 
             cardService.updateCard(card,card.cardID);
 
             newCardList.cardList.add(indexTo,card);
+
             listRepository.save(newCardList);
             listRepository.save(oldCardList);
